requirements.txt
advertools
google-cloud-translate
pytorch
numpy
pandas
scikit-learn
transformers